+++ srt-protocol/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod access;
 pub mod connection;
 pub mod listener;
 pub mod packet;
